requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
cryptography==2.4.2
Django==2.0.9
django-filter==2.0.0
djangorestframework==3.9.0
idna==2.8
pycparser==2.19
PyMySQL==0.9.2
pytz==2018.7
six==1.11.0
